+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added support for new level UI in the map tool
 - Added support for the new walkie-talkie item
 - Added support for the new scooters
+- Added support for the new cosmetics system
 
 **Changes**:
 - Left-handed mode now also swaps your controller bindings around
@@ -16,6 +17,9 @@
 **Removals**:
 - Removed F8 hotswapping completely
 - Removed support for v0.3.x
+
+**Fixes**:
+- Smooth turn should feel a bit smoother now when not in the center of your play area
 
 # 1.1.2